+++ src/FunFair.BuildCheck.ProjectChecks/Settings/OnlyExesShouldBePublishablePolicy.cs
@@ -10,7 +10,7 @@ using Microsoft.Extensions.Logging;
 namespace FunFair.BuildCheck.ProjectChecks.Settings
 {
     /// <summary>
-    ///     Checks that libraries are 'packable' except when they are test assemblies.
+    ///     Checks that exes are 'publishable' except when they are test assemblies.
     /// </summary>
     [SuppressMessage(category: "ReSharper", checkId: "ClassNeverInstantiated.Global", Justification = "Created by DI")]
     public sealed class OnlyExesShouldBePublishablePolicy : IProjectCheck